- upgraded to spice-gtk-0.11

+++ eclipse_projects/SpiceClient/jni/android-spice-src/common/draw.h
@@ -35,7 +35,7 @@
 #include <spice/enums.h>
 #include "mem.h"
 
-#define SPICE_GET_ADDRESS(addr) ((void *)(unsigned long)(addr))
+#define SPICE_GET_ADDRESS(addr) ((void *)(uintptr_t)(addr))
 #define SPICE_SET_ADDRESS(addr, val) ((addr) = (unsigned long)(val))
 
 typedef int32_t SPICE_FIXED28_4;
